Added seeder and select scripts to MySQL exercises.

+++ albums_migration.sql
@@ -7,7 +7,7 @@ create table albums (
 	artist VARCHAR(128) NOT NULL,-- string for storing the recording artist name
 	name VARCHAR(255) NOT NULL,-- string for storing a record name
 	release_date YEAR,-- integer representing year record was released (Note: MySQL also supports a YEAR data type!)
-	sales FLOAT(8,6),-- floating point value for number of records sold (in millions)
+	sales FLOAT(3,1),-- floating point value for number of records sold (in millions)
 	genre TEXT,
 	PRIMARY KEY (id)-- string for storing the record's genre(s)
 );
